updates to texture and textures class, buit mroe work doen on ttf class texture and textures can now take sdl_texture* directly, this is in build up to allow the entity system to also handle entities dedicated to showing text

+++ 2D top down/Entity.cpp
@@ -55,8 +55,8 @@ void Entity::MoveObjectTo(int x, int y)
 // The entities hold their direction as an angle, not as radian so we need to convert it back.
 void Object::MoveObjectAccoringToVel()
 {
-	point.x += cos(directionFacing * M_PI / 180) * velocity;
-	point.y += sin(directionFacing * M_PI / 180) * velocity;
+	point.x += (int)(cos(directionFacing * M_PI / 180) * velocity);
+	point.y += (int)(sin(directionFacing * M_PI / 180) * velocity);
 
 	if (point.x <= 0)
 		point.x = 0;
